Add Searchtab with base functionality
Added a Searchpage with a basic layout
Added new Activity for searching
Registered the Activity in the AndroidManifest.xml
Navigation to the Searchpage by Clicking on the third Icon in the Navbar
Added basic search functionality: You can search for the plantname. By default all plants are displayed. When searching all plants containing the name are displayed. When you finish the query by submitting, all fitting plants get shown. If there are no plants you'll see that there are no results for your query.